Refactor Knowledge Base upload functionality by removing the upload-from-url endpoint and associated validation schema. Update integration tests to reflect changes in the Knowledge Base structure and enhance error handling in the Pipeshub client for improved authorization header management.
