+++ backend/nodejs/apps/src/modules/knowledge_base/routes/kb.routes.ts
@@ -29,7 +29,6 @@ import {
   createNestedFolder,
   createRootFolder,
   uploadRecordsToKB,
-  uploadToKbUsingUrl,
   getKnowledgeHubNodes,
   moveRecord,
 } from '../controllers/kb_controllers';
@@ -57,7 +56,6 @@ import {
   getAllRecordsSchema,
   getAllKBRecordsSchema,
   uploadRecordsSchema,
-  uploadFromUrlSchema,
   uploadRecordsToFolderSchema,
   listKnowledgeBasesSchema,
   reindexRecordSchema,
@@ -445,15 +443,6 @@ export function createKnowledgeBaseRouter(
 
     // Upload handler
     uploadRecordsToKB(keyValueStoreService, appConfig, notificationService),
-  );
-
-  router.post(
-    '/:kbId/upload-from-url',
-    authMiddleware.authenticate,
-    requireScopes(OAuthScopeNames.KB_UPLOAD),
-    metricsMiddleware(container),
-    ValidationMiddleware.validate(uploadFromUrlSchema),
-    uploadToKbUsingUrl(keyValueStoreService, appConfig, notificationService),
   );
 
   // Upload records to a specific folder in the KB
