--- integration-tests/enterprise_search/.env.example
@@ -1,13 +1,0 @@
-# Environment variables for the enterprise_search integration tests.
-# Copy to integration-tests/.env.local and fill in the secrets.
-# Do not commit .env.local.
-
-# Node.js backend (Express API) base URL.
-PIPESHUB_BASE_URL=http://localhost:3000
-
-# Org user used to log in (initAuth -> authenticate) and own the seeded KB.
-PIPESHUB_TEST_USER_EMAIL=
-PIPESHUB_TEST_USER_PASSWORD=
-
-# Existing user id in the same org. Share/unshare tests are skipped when empty.
-PIPESHUB_TEST_SHARE_TARGET_USER_ID=
