+++ integration-tests/env.local.template
@@ -12,6 +12,9 @@ CLIENT_SECRET=
 PIPESHUB_TEST_USER_EMAIL=
 PIPESHUB_TEST_USER_PASSWORD=
 
+# Existing user id in the same org. Share/unshare tests skip when empty.
+PIPESHUB_TEST_SHARE_TARGET_USER_ID=
+
 
 TEST_GRAPH_DB_TYPE=neo4j   # arango / neo4j
