+++ integration-tests/env.prod.template
@@ -8,6 +8,9 @@ PIPESHUB_BASE_URL=https://test.pipeshub.com
 CLIENT_ID=
 CLIENT_SECRET=
 
+# Existing user id in the same org. Share/unshare tests skip when empty.
+PIPESHUB_TEST_SHARE_TARGET_USER_ID=
+
 # Neo4j Aura (driver uses TEST_NEO4J_* only)
 TEST_NEO4J_URI=
 TEST_NEO4J_USERNAME=
